// src/app/models/istoricCreateModel.ts
export interface IstoricCreateModel{
    numarLegitimatie:number;
    numeProba: string;
    numeCompetitie: string;
    categorie: string;
    locPesteToti:number;
    locPerGen: number;
    locPerCategorie: number;
    timpTotal: string;
    timpInot:string
    timpCiclism: string;
    timpAlergare: string;
    timpTranzit1: string;
    timpTranzit2: string;
    puncte: number;
}